making the dev2 more manageble
